binary clean parser
the content of files that git diffs might not be utf-8, and the
filepaths could also be weird encodings. we can't assume they are
decodable, so we have to treat them as binary.

python's json module will not marshal byte values by default, but we can
add that fairly easily. it will never marshal byte keys, but luckily our
binary handling decodes all the keys to ascii (which we know is safe)
beforehand.